Cosmetic fix.

+++ position_estimate.h
@@ -6,6 +6,6 @@
 
 using namespace std;
 
-const short cg_summa[] = {-5, -1, 0, 0, 1, 2, 5, 20, 30};
+const int cg_summa[] = {-5, -1, 0, 0, 1, 2, 5, 20, 30};
 
 Pos position_estimate(Field* field);
